一些优化，包括request的参数类型自由化，opcode变为int等

+++ src/main/java/com/peony/engine/framework/control/annotation/Request.java
@@ -20,5 +20,5 @@ public @interface Request {
      * opcode,框架根据该参数进行路径导航
      * TODO 这个可以考虑换成int类型
      * **/
-    short opcode();
+    int opcode();
 }
